requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
appdirs==1.4.4
asyncssh==2.22.0
atpublic==7.0.0
attrs==26.1.0
backports.functools-lru-cache==2.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
cachetools==7.0.5
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
databricks-sdk==0.102.0
deep-translator==1.11.4
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==1.1.0
dvc==3.67.1
dvc-data==3.18.3
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.136.0
filelock==3.28.0
flake8==7.3.0
Flask==3.1.3
flask-cors==6.0.2
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.4.0
gto==1.9.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.11.0
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
indic_transliteration==2.3.82
iniconfig==2.3.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
kombu==5.6.2
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mdurl==0.1.2
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.4
omegaconf==2.3.0
opentelemetry-api==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
packaging==26.1
pandas==2.3.3
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
prettytable==3.17.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
pydot==4.0.1
pyflakes==3.4.0
pygit2==1.19.2
Pygments==2.20.0
pygtrie==2.5.0
pyparsing==3.3.2
pytest==9.0.3
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
roman==5.2
ruamel.yaml==0.19.1
sacremoses==0.1.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
scmrepo==3.6.2
semver==3.0.4
sentencepiece==0.2.1
seqeval==1.2.2
setuptools==81.0.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
skops==0.13.0
smmap==5.0.3
soupsieve==2.8.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
sqltrie==0.11.2
starlette==1.0.0
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.14.0
torch==2.11.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.44.0
vine==5.1.0
voluptuous==0.16.0
waitress==3.0.2
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.8
yarl==1.23.0
zc.lockfile==4.0
zipp==3.23.1

